+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## <small>1.4.5 (2023-03-18)</small>
+## <small>1.4.6 (2023-03-22)</small>
 
+- 🐛 修复(pages): router ([2045786](https://github.com/2401345934/webpack-react-demo/commit/2045786))
 - init ([d7ddb87](https://github.com/2401345934/webpack-react-demo/commit/d7ddb87))
+- not auth plubsh ([cdfce05](https://github.com/2401345934/webpack-react-demo/commit/cdfce05))